Backend/Backend/requirements.txt
absl-py==2.0.0
asgiref==3.7.2
asttokens==2.0.8
astunparse==1.6.3
autopep8==2.0.2
backcall==0.2.0
bidict==0.22.1
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
cmake==3.18.2
colorama==0.4.5
contourpy==1.2.0
controller==0.1.0
cycler==0.12.1
debugpy==1.6.3
decorator==5.1.1
dill==0.3.7
Django==4.2.5
django-cors-headers==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
entrypoints==0.4
executing==0.10.0
flatbuffers==23.5.26
fonttools==4.44.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.2
h11==0.14.0
h5py==3.10.0
idna==3.4
ipykernel==6.15.1
ipython==8.4.0
jedi==0.18.1
joblib==1.3.2
jupyter-client==7.3.4
jupyter-core==4.11.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
nest-asyncio==1.5.5
numpy==1.26.1
oauthlib==3.2.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opt-einsum==3.3.0
packaging==21.3
pandas==2.1.3
parso==0.8.3
pbr==5.11.1
pickleshare==0.7.5
Pillow==10.1.0
ply==3.11
prompt-toolkit==3.0.30
protobuf==4.23.4
psutil==5.9.1
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pyFirmata==1.1.0
Pygments==2.13.0
pyparsing==3.0.9
pypdf==4.0.1
pyserial==3.5
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
PyWavelets==1.4.1
pywin32==304
pyzmq==23.2.1
qiskit==0.44.2
qiskit-terra==0.25.2.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
rustworkx==0.13.2
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.3
stack-data==0.4.0
stevedore==5.1.0
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tomli==2.0.1
tornado==6.2
traitlets==5.3.0
typing_extensions==4.8.0
tzdata==2022.6
urllib3==2.1.0
wcwidth==0.2.5
Werkzeug==3.0.1
wordcloud==1.9.3
wrapt==1.14.1
wsproto==1.2.0
